qcacld-3.0: Ensure phymode and chanwidth consistent
Propagation from qcacld-2.0 to qcacld-3.0.

After connecting to a candidate AP, the STA checks the
beacon's IE whether it needs to update channel width in
function __schBeaconProcessForSession. Currently, it uses
WMI_PEER_CHWIDTH to update the firmware parameter directly
and ignores Phymode. Firmware can only handle downgrade of
channel width and not upgrade. If Phymode is initially set
to VHT20 and later if channel width is to be changed to
VHT80. This has to be sent to firmware explicity via
PHYMODE WMI command otherwise firmware will ASSERT.

Fix is to send phymode along with channel width.

Change-Id: Id7405ab1c80bca8d7647b562af28f84b7501fd79
CRs-Fixed: 1052131