qcacld-3.0: Remove unnecessary assert in CDS timer callback
In the CDS timer callback, an attempt to post a message to the MC thread
can be made in cases where the MC thread is already stopped. In these
cases, the CDS timer callback will assert. Remove this assert, as
failing to post a message to the MC thread is expected during cases
such as shutdown or unload.

Change-Id: I018a792342cb31c9ead0f4782316afd56a48e67a
CRs-Fixed: 2041646