qcacld-3.0: Fix scan failure issue due to connection in progress
With QCA_WIFI_3_0_EMU set the def scan active cmd timeout is 10
minutes, which is very large to recover.

Also if scan cmd timeouts, abort_scan_indication is not set and
thus in case of scan for ssid, the connect command is not removed.

To fix this set scan cmd timeout to 30 sec and set abort scan
indication in scan cmd timeout.

Change-Id: Ieee450853958e221cf11a367dddb0b7e3367bf28
CRs-Fixed: 2044474