mfd: wcd9xxx-core: Synchronize codec down and reset calls from slimbus.
Sometimes after receiving the codec down callback from slimbus, the
context is back-scheduled and before it is complete, codec reset
callback is called and completed.The codec down context is complete
after this. This change of sequence causes codec reset pin to be
low after SSR. Ensure the sequence of codec down and codec reset
callbacks by waiting for codec down to complete in reset and using common
mutex for synchronization.

Change-Id: I2b9e119ce93a17017076738ab96d1fe4a1557982
Signed-off-by: Ramprasad Katkam <katkam@codeaurora.org>