qcacld-3.0: Remove tdls scan work queue
With current implementation if driver want to teardown tdls link,
it used to send a teardown indication to supplicant and supplicant
used to send teardown request to the peer. As this driver and
supplicant is going to take some time, driver is scheduling
workqueue for scan request to defer it for sometime until tdls links
gets teardown.

With this change, driver is going to send a deauth to peer directly
to avoid driver and supplicant interaction. So the need of scan work
queue can be eliminated.

CRs-Fixed: 2085541
Change-Id: If2aa32ced57cb9c9b154dfec7f0ba3c1d1f5f3af