qcacld-3.0: Add wait during module_init for static driver
Incase of the static driver there can be failure while accessing
the WCNSS_qcom_cfg.ini because ueventd has not enumerated the
file which results in probe failure.
Since the probe is happening in different context the failure will not
be updated to upper layers.

To mitigate the issue add wait for the static driver.

Change-Id: I90c8ec8a95e9f98ff420eb0fcf76a631da3ba0ba
CRs-Fixed: 2085748