qcacld-3.0: Revert the change for cleanup qdf list
wma_vdev_find_req will remove the request from queue if found it. If
remove it when peek it, it will abort to cleanup and memory leak. The
change for cleanup qdf list is for NULL mac_ctx case. Solution is
revert previous change and move mac_ctx checker back.

Change-Id: I06deedaf9f8a3c788621d8468e00d4d4852e2d9f
CRs-Fixed: 2090051