qcacld-3.0: cleanup scan queue in case of SSR
When the SSR and interface down happen in parallel,
the driver rejects the interface down since the
recovery is in progress. Kernel ignores the -EAGAIN
request from the driver and as part of NET_DOWN notification
in cfg80211_netdown_notifer the kernel invokes the
___cfg80211_scan_done to free the request but dose not send
scan_result indication, since it expects the scan_done work to
get scheduled and then broadcast the request to upperlayer.

According to the upstream commit f9d15d162b3acf28f85b3ac05c4883e5ed588d28
the reason the kernel doesnt broadcast the scan result to upperlayer
is that:
   1. interface starts being removed
   2. scan is cancelled by driver and cfg80211 is notified
   3. scan done work is scheduled
   4. interface is removed completely, rdev->scan_req is freed,
       event sent to userspace but scan done work remains pending
   5. new scan is requested on another virtual interface
   6. scan done work runs, freeing the still-running scan

To mitigate the issue, clean up the scan queue as part of netdev
going down and FW down event.

Change-Id: Id698044ebe30cf555300e909a916d74be1f8b94f
CRs-Fixed: 2115392