qcacld-3.0: Invoke callback based on sme command type
When total active scan commands exceed max scan
count, p2p remain on channel command is not added
to the sme command list. As part of cleanup, remain on
channel callback needs to be invoked but instead scan
callback is invoked which results in crash.

Fix is to invoke appropriate callback based on the
command type.

Change-Id: I1c58a0e1785e350c88a8a148d0f81d95aa4e2e98
CRs-Fixed: 2148688