qcacld-3.0: Allow Setkey in between scan on same vdev ID
For LFR 3.0, if firmware has an invalid PMKID, then host has to do
8-way handshake. So if a scan is already progress, the firmware
offloads 8-way handshake to host, the supplicant completes the 8-way
followed by 4-way HS and invokes the setkey command. In SME ,if scan is
pending in active scan command list ,then it holds the setkey command
on same vdev id , until the scan is complete. This causes the deauth
from AP as EAPOL timeout occurs due to m4 not sent by the firmware.

Fix is to allow the Setkey command on the same Vdev ID , even
when the Scan is in progress , i.e the scan command is still pending
in the smeScanCmdActive list.

Change-Id: Ia7723c12006e3d901109ab829cd35a91a8409578
CRs-Fixed: 2146124