qcacmn: Dont initialize gsonapi to napi
In monkey tests we have observed that gsonapi and NAPI instances
are same. During the creation of rx NAPI instances, the GSONAPI
corresponding to them are initialized to rx napi, and soon after
the right value is overwritten when the flush callbacks are
registered. However, durign this window, the same NAPI struct is
used by two different objects and therefore collisions happen.

Init the gsonapi with NULL until the real value is set later
during initialization.

Change-Id: Iaff3adf193a3eeb16603ce5e6d0e93ea6f456e52
CRs-Fixed: 2154636