qcacld-3.0: Dont initialize gsonapi to napi
In monkey tests we have observed that gsonapi and NAPI instances
are same. During the creation of rx NAPI instances, the GSONAPI
corresponding to them are initialized to rx napi, and soon after
the right value is overwritten when the flush callbacks are
registered. However, durign this window, the same NAPI struct is
used by two different objects and therefore collisions happen.

Init the gsonapi with NULL until the real value is set later
during initialization. Use proper NAPI on the kernel API.

Change-Id: Ifaaa31a0ab2427b5208e2c03c2bffa6beaffa40d
CRs-Fixed: 2159884