leds: qpnp-haptics: Allow duration for buffer mode
Currently, duration is not allowed to be set while playing the
pattern in buffer mode. This works fine. However when playing
multiple sample sets, the last sample set might be stopped early
as the play interrupt fires immediately after configuring the
new sample set. Fix this by calculating the duration based on the
wave samples configured and the play rate.

Change-Id: I67b1741d435e07bf7332f7e36231fefa9129a9a6
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: Shilpa Suresh <sbsure@codeaurora.org>