qcacld-3.0: Allocate and free memory for chan_info correctly
Memory for chan_info is allocated in hdd_wlan_startup() and
freed in hdd_wlan_exit(). When device boots up, memory gets
allocated for chan_info but if interface timer expires, all
the modules are closed and memory is checked for any leaks.
As memory for chan_info is freed in hdd_wlan_exit() so memory
for chan_info is still present and hence assert in
qdf_mem_check_for_leaks().

Allocate memory for chan_info in hdd_features_init() and free
it in hdd_features_deinit().

Change-Id: Ia8944da16a776d4de4fdfbb10b605ebd2b60f0b6
CRs-Fixed: 2178435