qcacld-3.0: Send ROAM_STOP if roam synch failed
if roam synch propagation fails, then a ROAM_STOP
command is sent to firmware so that the firmware
can send a HO_FAIL and hence gracefully disconnect
and connect again. But, in this case, the reason
code while sending ROAM_STOP was being sent as
REASON_ROAM_STOP_ALL and this was not letting the
command go to the firmware.

Send a ROAM_STOP command to firmware with an explicit
reason code for the failure so that it is not blocked
in WMA before sending it to firmware.

Change-Id: I4d7e2e525c145ca0e990dcef85948285e2186c63
CRs-Fixed: 2182671