qcacld-3.0: Match callback pointer type to callback function signature
Control Flow Integrity (CFI) implementation in the kernel.
CFI is an LLVM feature that protects function pointers against
control flow hijacking. Basically, when calling a function through
a pointer, the type of the pointer must match the signature
of the target function.

Hence, fix all the occurences of the function pointer type
which doesn't match the signature of the target function.

Change-Id: I194f2dc3c609ffda89cccaf09aa4f9ec955f0cd7
CRs-Fixed: 2196462