qcacld-3.0: Remove timer from sys msg queue when timer is stopped
In a scenario where timer has expired and the timer msg is posted
to the mc thread sys_msgq and from a different context the timer
has been stopped and freed, access of freed memory would occur
when mc thread resumes and processes the timer msg from sys_msgq.

Flush the timer msg pending in the sys_msgq when the timer is
stopped based on the unique timer_cookie in the timer msg.

Change-Id: I63af19d8782cde97af293d1b6143075f6c53cb31
CRs-Fixed: 2202177