qcacld-3.0: Refactor vendor OUI support in HOST driver
Presently, check for whether connection to a particular AP needs to be
initiated in 1x1 instead of 2x2 is presnt in LIM layer based on vendor
OUI and AP capabilities. This leads to mismatch between connection info
stored in CSR layer and LIM layer as the same info has not been indicated
back to CSR layer.

For some specific vendor APs, when connection is done in 11AC mode and at
later point of time, OMN frame is sent to AP for changing 2x2 to 1x1 due
to antenna sharing with LTE, AP is behaving abnormally as it is not able
to respond well to the OMN frame.

Check for initiating connection in 1x1 instead of 2x2 in CSR layer instead
of LIM layer to avoid capability mismatch.

Add extra check to initiate connection in 11N instead of 11AC for some
vendor APs based on their vendor OUIs and AP capabilities.

Change-Id: Ia1e451ea44367fa6d3d0c755b041013df99d7268
CRs-Fixed: 2224666