qcacld-3.0: Enable STA IPA offload for MDM platforms
In STA-SAP config, IPA offload is enabled only when first client is
connected to SAP. This is OK for MSM platforms. But for MDM platforms,
this is not the case. There's STA tethering cases, where traffic is
routed between STA and USB tethering devices while SAP is idle and no
wifi clients are connected to SAP. In such case, IPA offload needs to
be enabled for STA to promote performance.

Fix is to use available MDM_PLATFORM marco and STA IPA support is
decided upon driver build time.

Change-Id: If2bc016aa941f2c0651a2669f1169e631208326b
CRs-Fixed: 2068978