qcacld-3.0: Dont change the TDLS mode from eTDLS_SUPPORT_NOT_ENABLED
As part of concurrency tdls mode is changed to eTDLS_SUPPORT_NOT_ENABLED.
For p2p roc and hw mode change tdls mode will be changed to
eTDLS_SUPPORT_DISABLED and timers are started and in the timer handler
again the tdls mode is changed to eTDLS_SUPPORT_ENABLED without checking
the current mode. If the mode is eTDLS_SUPPORT_NOT_ENABLED then the mode
should not be changed to enabled as it can lead to connection tracker
enabling and wrong tdls mode in driver for concurrency.

Avoid calling the api wlan_hdd_tdls_set_mode if the current tdls mode is
eTDLS_SUPPORT_NOT_ENABLED.

Change-Id: If6b11a9deb41ecd4fd8c711cbdbbbe011d9308de
CRs-Fixed: 2253666