qcacld-3.0: Fill CSR session nss from MAC ctx nss for device mode
As part of refactoring the force NSS 1x1 WAR to CSR from LIM in
Ia1e451ea44367fa6d3d0c755b041013df99d7268, the PE session NSS filled
from the MAC context vdev_type_nss was overwritten by the NSS sent
from the CSR join request. In the CSR join request, the NSS was sent
as 1 if the AP required the force NSS 1x1 WAR and for all other APs
the NSS was sent as 0. This results in the PE session NSS to become
0 and the WMA sends the VDEV start request to FW with TX/RX preferred
streams as 1.

Fill the NSS from MAC context vdev_type_nss in the CSR join request
before sending the join request message to PE.

Change-Id: Ieec65a9369424f0cb64f21da5095fec2c4787214
CRs-Fixed: 2243159