qcacld-3.0: Increase SAP ACS scan priority to high from low
The function sap_goto_channel_sel triggers the pre start bss
scan for SAP. After this scan is queued, the hostapd process
gets scheduled after 3 secs and proceeds to select the channel
to start the SAP. If scan completion for the ACS scan was not
recieved, it selects the default channels. ACS scan is sent to
firmware with low priority unlike other scan.

Increasing the priority of the scan will ensure that the scan
completion is done prior to the other existing scans pending on
the queue.
So introduce a new scan_requestor_id macro ACS_SCAN_REQUESTOR_ID,
to distinguish the ACS scan and escalate the priority of this
scan to high.

Change-Id: Ibe558a4a323f276cce6eaabb3b62db217dbd5a94
CRs-Fixed: 2245200