qcacld-3.0: Set mScanPending to false only when no scans are pending
The flag mScanPending is used to keep a record if there are any scans
pending on the corresponding adapter or not. This flag is then referred
to when any abort scan operation is to be performed. Currently in the
failure case of processing a scan request in __iw_set_scan, the flag
mScanPending is set to false without verifying if any scans are still
pending. This results in the scans in queue not getting aborted during
unload/suspend resulting in a kernel panic.

Check if there are any pending scan requests before setting the flag
mScanPending to false.

Change-Id: I7de9c09b65dc3aed7656f3af9ff1ce5a33ba16b9
CRs-Fixed: 2248292