qcacld-3.0: Flush the restart sap work when the SAP is going down
Presently in the force switch to scc scenario to a channel which
sta is connected a work is scheduled to restart SAP. Inbetween
if the sap interface goes down the worker thread is not flushed.
So, when the worker thread gets scheduled it can access the freed
contexts of sap context.

Hence Flush the restart sap context when the sap is going down.

Change-Id: Ibb34392f33900aefb3085b06676818cbe9cbf066
CRs-Fixed: 2261585