qcacld-3.0: Choose channel from 2.4Ghz PCL list and ACS list as well
Presently, if a random channel needs to be selected, only 5Ghz channels
from PCL list are considered and if there is no valid 5Ghz channel in
PCL list then SAP is not started.

If the country domain is Japan, then the only valid 5Ghz channels are
channels 100-140 which are all DFS channels. Suppose, if a radar is
detected on any of this channel and if we consider only 5Ghz channels
in PCL list then SAP won't be started as there is no 5Ghz channel in
PCL list as PCL list does not contain DFS channels and DFS channel is
also not an option to start SAP again on after radar detection for the
next 30 minutes.

Consider 2.4Ghz PCL channels for starting SAP if there are no 5Ghz channels
in PCL list and consider ACS channel list as well if there are no 2.4Ghz or
5Ghz valid PCL channels present.

Change-Id: Icc67ce3158ae191f5275ce9ae23ef749ae72788e
CRs-Fixed: 2257146