qcacld-3.0: Remove redundant inclusion of wlan_hdd_apf.h
wlan_hdd_main.h and wlan_hdd_apf.h are included in each other
creating a circular dependancy. This can result in compilation
errors in some cases. The inclusion of the apf.h file in
main.h is redundant.

Remove the inclusion of apf.h from the main.h.

Change-Id: I7e419bc26d1a0150778ff075ab1f02741474a4a7
CRs-Fixed: 2267047