qcacld-3.0: Verify value before decrementing end_channel_idx
The local variable end_channel_idx in function hdd_ch_avoid_cb is
assigned value as per the available frequencies of different channels.
In the case the current channel frequency matches with an avoid
frequency range, the value is decremented by 1. When the value of
end_channel_idx is already 0, further decrementing it by one can lead to
out of bounds read.

Add a sanity check to assure that 0 is the minimum value that will be
assigned to end_channel_idx

Change-Id: I41b12c68b18271228083c79a060d7e6030ae356e
CRs-Fixed: 2260100