qcacld-3.0: Reduce scan time after Wi-Fi turned on in LONU driver
In LONU driver case, the first scan after wifi Off/On is
taking long time as DFS channels are getting included in
scan channel list. The reason for this DFS channel inclusion
is that the initial_scan_no_dfs_chnl is set to 0 after
1st scan and this parameter is not restored from ini on
Wifi OFF/ON as driver is not unloaded in case of LONU.

To address this issue, restore initial_scan_no_dfs_chnl
config param from ini on interface up(which is mapped to
Wifi On in case of LONU).

Change-Id: I07e872817433233bfbe68303e6cdfd05256c491a
CRs-Fixed: 2291478