qcacld-3.0: Fix TDLS Add Sta failure in HDD
After TDLS peer delete in lim, sme indicates the del sta response to
HDD. At that time staid in hdd_ctx is set to 0xFF as invalid. But,
In case of eCSR_ROAM_RESULT_DELETE_ALL_TDLS_PEER_IND indication to
HDD. The peer sta id is not set to 0xFF. This lead to improper
cleanup in HDD and failure in subsequent ADD sta requests.

Change-Id: I325d5993d2f1a75b84848983f6f94e264b149905
CRs-Fixed: 2298181