FROMLIST: ANDROID: binder: Add BINDER_GET_NODE_INFO_FOR_REF ioctl.
This allows the context manager to retrieve information about nodes
that it holds a reference to, such as the current number of
references to those nodes.

Such information can for example be used to determine whether the
servicemanager is the only process holding a reference to a node.
This information can then be passed on to the process holding the
node, which can in turn decide whether it wants to shut down to
reduce resource usage.

Change-Id: I5046aa020369483c5ed195b9daf522c23e0df6f8
Signed-off-by: Martijn Coenen <maco@android.com>
(cherry picked from commit e81cacbb05641b775f2639a3a9ab1a4f54422c0e)