qcacld-3.0: Possible memleak during con mode change
To change con mode, driver first wait until external threads
complete and then start con mode changing process.

In LL stats get request process, memory is allocated in request
and is freed in the response or when response timedout.

If LL stats get request is posted just after con mode waiting is
done for external threads to complete and LL stats is waiting for
its response, con mode handler will try to stop modules and will
result in memory leak as the memory for the request is not
released yet.

To address this issue, set load in progress to true in con mode
handler just before waiting for external threads to complete.

Change-Id: Ie3492c5609f99061a3441587c00a776321b5cc25
CRs-Fixed: 2321806