qcacld-3.0: Increase the pno wakelock timeout to 2 seconds
Currently wakelock timer WMA_PNO_MATCH_WAKE_LOCK_TIMEOUT
is of 1 sec which is resulting in multiple suspend/resume
after every second.There could be a possibility that once
the FW sends the NLO match and still scanning for the
remaining channels that could take more than 1 second and
since after 1 second driver is again receiving suspend and
sending same to FW which may result PNO scan completion
event to be delayed from FW because of which driver may
not send preferred network found indication.

To avoid this issue increase the pno wakelock timeout value
to 2 seconds allowing FW to complete the scan and to send
the pno completion event to host.

Change-Id: I601a318760a5843f4930bb56915ffe2461cdd12a
CRs-Fixed: 2324394