qcacld-3.0: Fetch the next adapter correctly during the system syspend
with Iacd2f97b01d7f6901d402908304a43c2c20d2380 while trying to flush
the ps timer logic is added to validate the session of the
adapter. There is a problem in the logic which will get struck in the 
infinite loop.

Fetch the next adapter when the session of the adapter is invalid.

Change-Id: I68c564b7cabe72c2adc0548ace55011e16282c1b
CRs-Fixed: 2349172