qcacld-3.0: Don't close SAP session internally
SAP session is getting closed internally during stop_bss,
there might be some commands pending to be processed.
During del virtual interface, in stop adapter there is a wait
for sme close session, since this session is getting closed
in stop_bss, stop_adapter will not wait for the pending
commands to be processed and del virtual interface will
eventually delete the interface and will close the adapter.
Now if the pending commands will get the chance to execute,
it will try to access the deleted adapter which is already
deleted and eventually leads to a NULL pointer access.

To prevent this, do not close the SAP session during stop bss,
hdd should take of closing the session during stop adapter.

Change-Id: I54af625b89021b8ccf113d9c4d134d5eec47080a
CRs-Fixed: 2354943